% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{package}
\name{dtscan-package}
\alias{dtscan-package}
\title{dtscan: Triangle‑Based Density Clustering of 2D Points}
\description{
The dtscan package provides R bindings to a high‑performance Rust
implementation of triangle‑based dtscan via Delaunay triangulation.
}
\details{
Clusters emerge by iteratively linking points whose connecting edges fall
below a user‑defined distance threshold, effectively applying a
dtscan‑style density rule on the triangulation graph.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/valcu/dtscan}
  \item Report bugs at \url{https://github.com/valcu/dtscan/issues}
}

}
\author{
\strong{Maintainer}: Mihai Valcu \email{mvalcu@gwdg.de} (\href{https://orcid.org/0000-0002-6907-7802}{ORCID})

}
\keyword{internal}
